refactor: move ModelWrapper class to own file
